Extract configurable menu bar controller
